requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-ratelimit==4.1.0
django-tailwind==4.2.0
djangorestframework==3.16.0
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
qrcode==8.2
redis==6.2.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
